+++ src/index.js
@@ -1,10 +1,17 @@
 import { VivViewerLayer, StaticImageLayer } from './layers';
-import VivViewer from './VivViewer';
+import { VivViewer, PictureInPictureViewer, SideBySideViewer } from './viewers';
+import { VivView, OverviewView, DetailView, SideBySideView } from './views';
 import { createTiffPyramid, createZarrLoader, ZarrLoader } from './loaders';
 
 export {
   VivViewerLayer,
   VivViewer,
+  VivView,
+  OverviewView,
+  PictureInPictureViewer,
+  SideBySideView,
+  SideBySideViewer,
+  DetailView,
   StaticImageLayer,
   ZarrLoader,
   createTiffPyramid,