+++ src/layers/VivViewerLayer/VivViewerLayer.js
@@ -3,6 +3,22 @@ import VivViewerLayerBase from './VivViewerLayerBase';
 import StaticImageLayer from '../StaticImageLayer';
 import { padColorsAndSliders } from '../utils';
 
+/**
+ * This layer generates a VivViewerLayer (tiled) and a StaticImageLayer (background for the tiled layer)
+ * @param {Object} props
+ * @param {Array} props.sliderValues List of [begin, end] values to control each channel's ramp function.
+ * @param {Array} props.colorValues List of [r, g, b] values for each channel.
+ * @param {Array} props.channelIsOn List of boolean values for each channel for whether or not it is visible.
+ * @param {number} props.opacity Opacity of the layer.
+ * @param {string} props.colormap String indicating a colormap (default: '').  The full list of options is here: https://github.com/glslify/glsl-colormap#glsl-colormap
+ * @param {Array} props.domain Override for the possible max/min values (i.e something different than 65535 for uint16/'<u2').
+ * @param {string} props.viewportId Id for the current view.
+ * @param {Object} props.loader Loader to be used for fetching data.  It must implement/return `getTile`, `dtype`, `numLevels`, and `tileSize`, and `getRaster`.
+ * @param {Array} props.loaderSelection Selection to be used for fetching data.
+ * @param {String} props.id Unique identifier for this layer.
+ * @param {String} props.onTileError Custom override for handle tile fetching errors.
+ */
+
 export default class VivViewerLayer extends CompositeLayer {
   renderLayers() {
     const {