+++ src/viewers/index.js
@@ -1,0 +1,3 @@
+export { default as VivViewer } from './VivViewer';
+export { default as PictureInPictureViewer } from './PictureInPictureViewer';
+export { default as SideBySideViewer } from './SideBySideViewer';