+++ src/views/index.js
@@ -1,0 +1,4 @@
+export { default as DetailView } from './DetailView';
+export { default as OverviewView } from './OverviewView';
+export { default as VivView } from './VivView';
+export { default as SideBySideView } from './SideBySideView';