+++ src/views/utils.js
@@ -1,0 +1,25 @@
+import { OrthographicView } from '@deck.gl/core';
+
+export function getVivId(id) {
+  return `-#${id}#`;
+}
+/**
+ * Create a boudning box from a viewport based on passed-in viewState.
+ * @param {viewState} Object The viewState for a certain viewport.
+ * @returns {View} The DeckGL View for this viewport.
+ */
+export function makeBoundingBox(viewState) {
+  const viewport = new OrthographicView().makeViewport({
+    // From the current `detail` viewState, we need its projection matrix (actually the inverse).
+    viewState,
+    height: viewState.height,
+    width: viewState.width
+  });
+  // Use the inverse of the projection matrix to map screen to the view space.
+  return [
+    viewport.unproject([0, 0]),
+    viewport.unproject([viewport.width, 0]),
+    viewport.unproject([viewport.width, viewport.height]),
+    viewport.unproject([0, viewport.height])
+  ];
+}