+++ CHANGELOG.md
@@ -4,7 +4,12 @@
 
 ### Added
 
+- New classes for views/viewers
+- Zoom/pan lock buttons and linked views in demo
+
 ### Changed
+
+- Refactored VivViewer to take in arbitrary views/layers
 
 ## 0.1.5